Rewrote attack and takeDamage functions; wrote actor presets for orc, troll, and player in the createActor function in the actor builder; changed combat system to work with dice instead of fixed damage values; changed some functions in actor builder into templates for future expansion; combat text will check for proper names and add *the* when necessary; the engine now has a static member for rng; engine has static class with direct access to messages in console rather than through m_gui; added messageFirstUpper static method to engine which ensures the message will start with a capital letter; started changing text colors to namespaced globals to ease future change; deleted YN enums in actor builder; started removing comment separators; started with a a glimpse of evasion in combat; left targeting system imcomplete.

+++ Gui.cpp
@@ -42,7 +42,7 @@ void Gui::render()
     PlayerAi *p_ai{dynamic_cast<PlayerAi *>(Engine::s_engine->m_player->m_ai.get())};
     char xpTxt[128];
     sprintf(xpTxt, "XP(%d)", p_ai->m_xp_lvl);
-    renderBar(1, 5, BAR_WIDTH, xpTxt, Engine::s_engine->m_player->m_destructible->m_xp, p_ai->getNextLevel(), TCODColor::lightViolet, TCODColor::darkViolet);
+    renderBar(1, 5, BAR_WIDTH, xpTxt, (float)Engine::s_engine->m_player->m_destructible->m_xp, (float)p_ai->getNextLevel(), TCODColor::lightViolet, TCODColor::darkViolet);
 
     // Draw the message log
     int y{1};
